+++ src/sidenav/components/SidenavGroup.vue
@@ -6,6 +6,7 @@ export default connect({
   gettersToProps: {
     groupId: 'currentGroup/id',
     wallUnreadCount: 'currentGroup/conversationUnreadCount',
+    pendingApplications: 'applications/forCurrentGroupPending',
   },
 })('SidenavGroup', SidenavGroupUI)
 </script>