+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Please document your changes in this format:
 - Invitations can be resent after 1 hour #837 @mvellasco
 - Move place description into modal @tiltec
 - Move notification settings to user settings @tiltec @djahnie
+- Move applications from group wall into sidenav @tiltec
 
 ### Fixed
 - Permission checks in main routes had a typo and were not active @djahnie @tiltec